# Conflicts:
#	OpenXLSX/external/zippy/zippy.hpp
#	OpenXLSX/headers/XLComments.hpp
#	OpenXLSX/headers/XLFormula.hpp
#	OpenXLSX/headers/XLSheet.hpp
#	OpenXLSX/headers/XLXmlFile.hpp
#	OpenXLSX/headers/XLXmlParser.hpp
#	OpenXLSX/sources/XLSheet.cpp

+++ OpenXLSX/sources/XLZipArchive.cpp
@@ -44,10 +44,33 @@ YM      M9  MM    MM MM       MM    MM   d'  `MM.    MM            MM   d'  `MM.
  */
 
 // ===== External Includes ===== //
-#include <zippy.hpp>
+#include <string>   // std::string
+
+#ifdef USE_LIBZIP
+	#include "detail/LibZip.hpp"
+	#define ZipLibraryName_impl    LibZip::ZipLibraryName
+	#define ZipLibraryVersion_impl LibZip::ZipLibraryVersion
+#else
+	#include "detail/Zippy.hpp"
+	#define ZipLibraryName_impl    Zippy::ZipLibraryName
+	#define ZipLibraryVersion_impl Zippy::ZipLibraryVersion
+#endif
 
 // ===== OpenXLSX Includes ===== //
 #include "XLZipArchive.hpp"
+#include "XLException.hpp"
+
+namespace OpenXLSX {
+    /**
+     * @details
+     */
+    const char *ZipLibraryName() { return ZipLibraryName_impl(); }
+
+    /**
+     * @details
+     */
+    const char *ZipLibraryVersion() { return ZipLibraryVersion_impl(); };
+}
 
 using namespace OpenXLSX;
 
@@ -57,9 +80,53 @@ using namespace OpenXLSX;
 XLZipArchive::XLZipArchive() : m_archive(nullptr) {}
 
 /**
+ * @details CAUTION: shallow copy (explicit default constructor)
+ */
+XLZipArchive::XLZipArchive(const XLZipArchive& other)
+ : m_archive(other.m_archive)
+{}
+
+/**
+ * @details
+ */
+XLZipArchive::XLZipArchive(XLZipArchive&& other) noexcept
+ : m_archive(std::move(other.m_archive))
+{}
+
+/**
  * @details
  */
 XLZipArchive::~XLZipArchive() = default;
+
+// /**
+//  * @details CAUTION: shallow copy (explicit default copy assignment)
+//  */
+// XLZipArchive& XLZipArchive::operator=(const XLZipArchive& other)
+// {
+//     m_archive = other.m_archive;
+//     return *this;
+// }
+
+/**
+ * @details
+ */
+XLZipArchive& XLZipArchive::operator=(XLZipArchive&& other) noexcept
+{
+    m_archive = std::move(other.m_archive);
+	 return *this;
+}
+
+/**
+ * @details
+ */
+bool XLZipArchive::usesZippy()
+{
+#ifdef USE_LIBZIP
+    return false;
+#else
+    return true;
+#endif
+}
 
 /**
  * @details
@@ -84,7 +151,8 @@ bool XLZipArchive::isOpen() const
  */
 void XLZipArchive::open(const std::string& fileName)
 {
-    m_archive = std::make_shared<Zippy::ZipArchive>();
+    if (isOpen()) throw XLInputError("XLZipArchive::open: archive is already open"); // prevent double open
+    m_archive = std::make_shared<XLZipImplementation>();
     try {
         m_archive->Open(fileName);
     }
@@ -99,6 +167,7 @@ void XLZipArchive::open(const std::string& fileName)
  */
 void XLZipArchive::close()
 {
+    if (!m_archive) throw XLInputError("XLZipArchive::close: archive is not open"); // prevent SEGFAULT
     m_archive->Close();
     m_archive.reset();
 }
@@ -106,8 +175,22 @@ void XLZipArchive::close()
 /**
  * @details
  */
+void XLZipArchive::commitChanges()
+{
+    if (!m_archive) throw XLInputError("XLZipArchive::commitChanges: archive is not open"); // prevent SEGFAULT
+#ifdef USE_LIBZIP
+    m_archive->CommitChanges(); // libzip needs a zip archive close & reopen
+#else
+    // no-op for zippy
+#endif
+}
+
+/**
+ * @details
+ */
 void XLZipArchive::save(const std::string& path) // NOLINT
 {
+    if (!m_archive) throw XLInputError("XLZipArchive::save: archive is not open"); // prevent SEGFAULT
     m_archive->Save(path);
 }
 
@@ -116,7 +199,17 @@ void XLZipArchive::save(const std::string& path) // NOLINT
  */
 void XLZipArchive::addEntry(const std::string& name, const std::string& data) // NOLINT
 {
+    if (!m_archive) throw XLInputError("XLZipArchive::addEntry: archive is not open"); // prevent SEGFAULT
     m_archive->AddEntry(name, data);
+}
+
+/**
+ * @details
+ */
+void XLZipArchive::addEntryAndCommit(const std::string& name, const std::string& data) // NOLINT
+{
+    addEntry(name, data);
+    commitChanges();       // ensure that entry name is directly available for calls to getEntry
 }
 
 /**
@@ -124,6 +217,7 @@ void XLZipArchive::addEntry(const std::string& name, const std::string& data) //
  */
 void XLZipArchive::deleteEntry(const std::string& entryName) // NOLINT
 {
+    if (!m_archive) throw XLInputError("XLZipArchive::deleteEntry: archive is not open"); // prevent SEGFAULT
     m_archive->DeleteEntry(entryName);
 }
 
@@ -131,12 +225,42 @@ void XLZipArchive::deleteEntry(const std::string& entryName) // NOLINT
  * @details
  */
 std::string XLZipArchive::getEntry(const std::string& name) const {
+    if (!m_archive) throw XLInputError("XLZipArchive::getEntry: archive is not open"); // prevent SEGFAULT
+#ifdef USE_LIBZIP
+    return m_archive->GetEntryDataAsString(name);
+#else
     return m_archive->GetEntry(name).GetDataAsString();
+#endif
 }
 
 /**
  * @details
  */
 bool XLZipArchive::hasEntry(const std::string& entryName) const {
+    if (!m_archive) throw XLInputError("XLZipArchive::hasEntry: archive is not open"); // prevent SEGFAULT
     return m_archive->HasEntry(entryName);
+}
+
+/**
+ * @details
+ */
+ssize_t XLZipArchive::entryCount() const
+{
+#ifdef USE_LIBZIP
+    return m_archive->EntryCount();
+#else
+    return 0; // TODO: can Zippy support this?
+#endif
+}
+
+/**
+ * @details
+ */
+std::string XLZipArchive::entryName( int index ) const
+{
+#ifdef USE_LIBZIP
+    return m_archive->EntryName(index);
+#else
+    return ""; // TODO: can Zippy support this?
+#endif
 }
